- Add protection to StandardPacker to prevent segmentation faults from a NULL    pointer dereference in the case of corrupted reference information, due    to truncation when attempting to pack it to a 32 bit int. Fixes issues    seen when running over Stripping21 MC samples where some Track references    in packed ProtoParticles have been corrupted due to data truncation. In    these cases the smart refs are now just skipped (so set to NULL).  - Update the ProtoParticle packing to include versioning and define a new    version 1 that stores all SmartRefs as packed 64 bit ints, to avoid    truncation that can occur when using only 32 bit ints.  - ToDo - The (poorly thought out) use of 32bit packed references should all           be migrated to use 64 bit packing instead, to finally prevent this           sort of issue happening again.
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@184970 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 3def82944cc86e32a6963b102d5fbc5fdb8cf7ed

+++ LHCb/Event/EventPacker/src/component/ReadPackedDst.cpp
@@ -195,7 +195,7 @@ StatusCode ReadPackedDst::execute() {
       put( protos, name + m_postfix );
       processLinks( protos, version );
       getFromBlob<LHCb::PackedProtoParticle> ( protos->protos() , blobs );
-      getFromBlob<int>                       ( protos->refs()   , blobs );
+      getFromBlob<long long>                 ( protos->refs()   , blobs );
       getFromBlob<std::pair<int,int> >       ( protos->extras() , blobs );
 
     } else if ( LHCb::CLID_PackedRecVertices     == classID ) {
